Finised user sign in system and autorization system.
No longer using cookies to handle autorization as they did not work as well with Blazor Server.

Using a claims based system. it logs out users if they refresh the page but that problem can be looked at after everything else is working and in a state I am happy with.