Pokedex page is mostly complete.
page size customisation, and pager functionalities working.

I would liek to make it so if you click anywhere on the card it brings you to that pokemons page but that can wait.

Working on converting PokeApi Information to my base class.

lots of complexity i would i liek to add like forms or eveolution chains but that can wait for teh very end.

need to actuly start adding comments.